+++ mempool/reactor.go
@@ -28,7 +28,6 @@ type MempoolReactor struct {
 	p2p.BaseReactor
 	config  *cfg.MempoolConfig
 	Mempool *Mempool
-	evsw    types.EventSwitch
 }
 
 // NewMempoolReactor returns a new MempoolReactor with the given config and mempool.
@@ -148,11 +147,6 @@ func (memR *MempoolReactor) broadcastTxRoutine(peer Peer) {
 		next = next.NextWait()
 		continue
 	}
-}
-
-// SetEventSwitch implements events.Eventable.
-func (memR *MempoolReactor) SetEventSwitch(evsw types.EventSwitch) {
-	memR.evsw = evsw
 }
 
 //-----------------------------------------------------------------------------