+++ rpc/core/pipe.go
@@ -36,7 +36,6 @@ type P2P interface {
 
 var (
 	// external, thread safe interfaces
-	eventSwitch   types.EventSwitch
 	proxyAppQuery proxy.AppConnQuery
 
 	// interfaces defined in types and above
@@ -51,13 +50,10 @@ var (
 	addrBook         *p2p.AddrBook
 	txIndexer        txindex.TxIndexer
 	consensusReactor *consensus.ConsensusReactor
+	eventBus         *types.EventBus
 
 	logger log.Logger
 )
-
-func SetEventSwitch(evsw types.EventSwitch) {
-	eventSwitch = evsw
-}
 
 func SetBlockStore(bs types.BlockStore) {
 	blockStore = bs
@@ -101,4 +97,8 @@ func SetConsensusReactor(conR *consensus.ConsensusReactor) {
 
 func SetLogger(l log.Logger) {
 	logger = l
+}
+
+func SetEventBus(b *types.EventBus) {
+	eventBus = b
 }